grupo1/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.23.2
python-dotenv==1.0.1
requests==2.32.0